backend/requirements.txt
fastapi==0.95.0
requests==2.28.2
pydantic==1.10.7
pillow==9.5.0
python-dotenv==1.0.0
openai==0.27.0
